Add default RF allocation.

+++ interfaces/IRewardFlow.sol
@@ -7,7 +7,7 @@ interface IRewardFlow {
     function rfFactory() external view returns(address);
     function payForward() external returns(address, uint);
     function setArtifact() external;
-    function submitAllocation(address targetAddr, uint allocAmt) external returns(uint);
+    function submitAllocation(address targetAddr, uint8 allocAmt) external returns(uint);
     function redeemReward(address claimer, uint redeemAmt) external returns (uint);
     event Allocate(address indexed from, address indexed to, uint256 value);
     event Distribute(address indexed from, address indexed to, uint256 value);
